--- a/src/Symfony/Bridge/Doctrine/Form/Type/EntityType.php
+++ b/src/Symfony/Bridge/Doctrine/Form/Type/EntityType.php
@@ -82,7 +82,7 @@ class EntityType extends DoctrineType
 
         return [
             $queryBuilder->getQuery()->getSQL(),
-            array_map([$this, 'parameterToArray'], $queryBuilder->getParameters()->toArray()),
+            array_map($this->parameterToArray(...), $queryBuilder->getParameters()->toArray()),
         ];
     }