--- a/src/Symfony/Bridge/Twig/Extension/AssetExtension.php
+++ b/src/Symfony/Bridge/Twig/Extension/AssetExtension.php
@@ -35,8 +35,8 @@ final class AssetExtension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('asset', [$this, 'getAssetUrl']),
-            new TwigFunction('asset_version', [$this, 'getAssetVersion']),
+            new TwigFunction('asset', $this->getAssetUrl(...)),
+            new TwigFunction('asset_version', $this->getAssetVersion(...)),
         ];
     }