--- a/src/Symfony/Bridge/Twig/Extension/HttpFoundationExtension.php
+++ b/src/Symfony/Bridge/Twig/Extension/HttpFoundationExtension.php
@@ -36,8 +36,8 @@ final class HttpFoundationExtension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('absolute_url', [$this, 'generateAbsoluteUrl']),
-            new TwigFunction('relative_path', [$this, 'generateRelativePath']),
+            new TwigFunction('absolute_url', $this->generateAbsoluteUrl(...)),
+            new TwigFunction('relative_path', $this->generateRelativePath(...)),
         ];
     }