--- a/src/Symfony/Bridge/Twig/Extension/WebLinkExtension.php
+++ b/src/Symfony/Bridge/Twig/Extension/WebLinkExtension.php
@@ -37,12 +37,12 @@ final class WebLinkExtension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('link', [$this, 'link']),
-            new TwigFunction('preload', [$this, 'preload']),
-            new TwigFunction('dns_prefetch', [$this, 'dnsPrefetch']),
-            new TwigFunction('preconnect', [$this, 'preconnect']),
-            new TwigFunction('prefetch', [$this, 'prefetch']),
-            new TwigFunction('prerender', [$this, 'prerender']),
+            new TwigFunction('link', $this->link(...)),
+            new TwigFunction('preload', $this->preload(...)),
+            new TwigFunction('dns_prefetch', $this->dnsPrefetch(...)),
+            new TwigFunction('preconnect', $this->preconnect(...)),
+            new TwigFunction('prefetch', $this->prefetch(...)),
+            new TwigFunction('prerender', $this->prerender(...)),
         ];
     }