--- a/src/Symfony/Bundle/FrameworkBundle/Command/ConfigDebugCommand.php
+++ b/src/Symfony/Bundle/FrameworkBundle/Command/ConfigDebugCommand.php
@@ -127,7 +127,6 @@ EOF
         $kernel->boot();
 
         $method = new \ReflectionMethod($kernel, 'buildContainer');
-        $method->setAccessible(true);
         $container = $method->invoke($kernel);
         $container->getCompiler()->compile($container);