… .appveyor.yml
… .github/workflows/integration-tests.yml
… .github/workflows/intl-data-tests.yml
… .github/workflows/psalm.yml
… .github/workflows/unit-tests.yml
… composer.json
… src/Symfony/Bridge/Doctrine/Tests/DependencyInjection/DoctrineExtensionTest.php
… src/Symfony/Bridge/Doctrine/Tests/Form/Type/EntityTypeTest.php
… src/Symfony/Bridge/Doctrine/Tests/PropertyInfo/DoctrineExtractorTest.php
… src/Symfony/Bridge/Doctrine/Tests/Validator/DoctrineLoaderTest.php
… src/Symfony/Bridge/Doctrine/composer.json
… src/Symfony/Bridge/Monolog/composer.json
… src/Symfony/Bridge/ProxyManager/composer.json
… src/Symfony/Bridge/Twig/Tests/Translation/TwigExtractorTest.php
… src/Symfony/Bridge/Twig/composer.json
… src/Symfony/Bundle/DebugBundle/composer.json
… src/Symfony/Bundle/FrameworkBundle/Routing/Router.php
… src/Symfony/Bundle/FrameworkBundle/Tests/CacheWarmer/AnnotationsCacheWarmerTest.php
… src/Symfony/Bundle/FrameworkBundle/Tests/Command/CachePoolClearCommandTest.php
… src/Symfony/Bundle/FrameworkBundle/Tests/Command/CachePoolDeleteCommandTest.php
… src/Symfony/Bundle/FrameworkBundle/Tests/Command/CachePruneCommandTest.php
… src/Symfony/Bundle/FrameworkBundle/Tests/Console/Descriptor/AbstractDescriptorTest.php
… src/Symfony/Bundle/FrameworkBundle/Tests/Console/Descriptor/ObjectsProvider.php
… src/Symfony/Bundle/FrameworkBundle/Tests/Functional/CachePoolClearCommandTest.php
… src/Symfony/Bundle/FrameworkBundle/composer.json
… src/Symfony/Bundle/SecurityBundle/Tests/Functional/Bundle/RememberMeBundle/Security/StaticTokenProvider.php
… src/Symfony/Bundle/SecurityBundle/composer.json
… src/Symfony/Bundle/TwigBundle/Tests/DependencyInjection/TwigExtensionTest.php
… src/Symfony/Bundle/TwigBundle/composer.json
… src/Symfony/Bundle/WebProfilerBundle/Tests/EventListener/WebDebugToolbarListenerTest.php