@94noni 2022-02-09
According to:

- https://packagist.org/packages/symfony/console/php-stats#6
- https://packagist.org/packages/symfony/http-foundation/php-stats#6

~50% at the time of writing, is already running php v8.1 :)