@jenkoian 2022-02-09
Not sure if this will affect your decision or not but [AWS EB doesn't yet support PHP 8.1](https://docs.aws.amazon.com/elasticbeanstalk/latest/platforms/platform-history-php.html).  I think most of the other PAAS providers offer PHP 8.1 though (except Google AppEngine which afaict doesn't even offer 8.0 support yet).