@nicolas-grekas 2022-02-09
>  ldap_rename(): Passing null to parameter #4 ($new_parent) of type string is deprecated

This shoud be fixed on lower branches. PR welcome anyone, that's a good first contrib :)

> TypeError: Memcached::setMulti(): Argument #2 ($expiration) must be of type int, int given

:facepalm: 

Any way, this PR is ready to go from a technical pov, remaining failures are false positives.