@nicolas-grekas 2022-02-09
> Not sure if this will affect your decision or not but [AWS EB doesn't yet support PHP 8.1](https://docs.aws.amazon.com/elasticbeanstalk/latest/platforms/platform-history-php.html)

Thanks for the info @jenkoian. That shouldn't be a blocker to me.