@derrabus 2022-02-09
The Packagist statistics seem to be in your favor:

* Among the consumers of FrameworkBundle 6.0, PHP 8.1 is already the dominant PHP branch. https://packagist.org/packages/symfony/framework-bundle/php-stats#6.0
* If we look at the 4.4 branch, our LTS that is approaching EOL, we see that we maintain support for a PHP version 7.1 that is almost irrelevant for the consumers of that package. https://packagist.org/packages/symfony/framework-bundle/php-stats#4.4
* On 5.4, it's a similar story. We support PHP 7.2 and 7.3 for a very small fraction of our userbase: https://packagist.org/packages/symfony/framework-bundle/php-stats#5.4

However, this is a major change in our release strategy: we haven't bumped the PHP language level in a minor release yet. If we take that road, I believe we should announce this a bit ahead of time. The fixed release cycles of PHP and Symfony should encourage development teams to schedule their framework and infrstructure upgrades ahead of time. Projects that chose to start with or upgrade to Symfony 6 might have done so trusting that they can remain on PHP 8.0 for a certain period of time. If we bump now, we basically force those projects to either reschedule infrastructure upgrades or to remain on Symfony 6.0 beyond its EOL.

If we would've discussed and announced that course of action before the 6.0 release, I would be all cool with it. But bumping to PHP 8.1 on such a short notice might make Symfony's release strategy appear erradic.

How would you feel about postponing this language level bump to the 6.2 release? That could give downstream projects a bit more time to adjust their development schedule accordingly.