@nicolas-grekas 2022-02-09
It's important to acknowledge the reason why I'm proposing this. We *do* have a strong case here, and it's not going to go away. PHP 8.0+preloading+typed props is broken, see linked issue.

I'm personally fine postponing to 6.2 if that's useful for the ecosystem. We'll need to reverse once more the types added to private properties before tagging 6.1.

On the other hand, we are months ahead the release of 6.1. we're not rushing anything here.