@xjm 2022-02-09
Just following up on @catch56's comment -- Would it be possible to let us know by *March 17, 2022* whether Symfony 6 is likely to introduce a PHP 8.1 requirement (whether in 6.1, 6.2, or some future release)? That's the deadline we committed to for announcing any further increase to Drupal 10's platform requirements. (We try to only increase PHP version requirements in majors because our audience includes a lot of long-tail end users who struggle with the disruption of it; when we dropped PHP 5 support in a minor release we gave like 18 months' advance warning and folks still struggled.) 

We'd really appreciate the favor if it's possible. :) We'll be watching this PR closely. Thanks! 🤞