@GrahamCampbell 2022-02-09
For Laravel 9 users, extending the security EOL of the 6.0.x series to only the EOL of the 6.1.x series wouldn't be enough, but delaying the bump to 6.2, and extending the life of 6.1 to EOL of 6.2 would work I think.