@fbourigault 2022-02-10
With `ab -n 10000 -c 5 127.0.0.1:8004/fr/blog/`, preload seems ~7% slower!

Disclamer: this bench was run on my laptop. The results may have been altered by other running programs.

With preload:
```
This is ApacheBench, Version 2.3 <$Revision: 1843412 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 1000 requests
Completed 2000 requests
Completed 3000 requests
Completed 4000 requests
Completed 5000 requests
Completed 6000 requests
Completed 7000 requests
Completed 8000 requests
Completed 9000 requests
Completed 10000 requests
Finished 10000 requests


Server Software:        
Server Hostname:        127.0.0.1
Server Port:            8004
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128
Server Temp Key:        ECDH P-256 256 bits

Document Path:          /fr/blog/
Document Length:        28066 bytes

Concurrency Level:      5
Time taken for tests:   45.488 seconds
Complete requests:      10000
Failed requests:        6
   (Connect: 0, Receive: 0, Length: 6, Exceptions: 0)
Non-2xx responses:      6
Total transferred:      282681094 bytes
HTML transferred:       280491604 bytes
Requests per second:    219.84 [#/sec] (mean)
Time per request:       22.744 [ms] (mean)
Time per request:       4.549 [ms] (mean, across all concurrent requests)
Transfer rate:          6068.75 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        3    5   0.9      5      22
Processing:     3   18   3.6     18     111
Waiting:        2   18   3.6     18     110
Total:          7   23   3.8     23     115

Percentage of the requests served within a certain time (ms)
  50%     23
  66%     24
  75%     24
  80%     25
  90%     25
  95%     26
  98%     28
  99%     29
 100%    115 (longest request)
```

Without preload:
```
This is ApacheBench, Version 2.3 <$Revision: 1843412 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 1000 requests
Completed 2000 requests
Completed 3000 requests
Completed 4000 requests
Completed 5000 requests
Completed 6000 requests
Completed 7000 requests
Completed 8000 requests
Completed 9000 requests
Completed 10000 requests
Finished 10000 requests


Server Software:        
Server Hostname:        127.0.0.1
Server Port:            8004
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128
Server Temp Key:        ECDH P-256 256 bits

Document Path:          /fr/blog/
Document Length:        28066 bytes

Concurrency Level:      5
Time taken for tests:   48.771 seconds
Complete requests:      10000
Failed requests:        0
Total transferred:      282850000 bytes
HTML transferred:       280660000 bytes
Requests per second:    205.04 [#/sec] (mean)
Time per request:       24.386 [ms] (mean)
Time per request:       4.877 [ms] (mean, across all concurrent requests)
Transfer rate:          5663.57 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        3    5   0.9      5      18
Processing:    13   20   6.0     20     306
Waiting:       13   19   6.0     19     305
Total:         17   24   6.2     24     309

Percentage of the requests served within a certain time (ms)
  50%     24
  66%     25
  75%     26
  80%     26
  90%     28
  95%     28
  98%     30
  99%     31
 100%    309 (longest request)
```